✨ CRUDに関する実装

+++ backend/src/memo/memo.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { Memo } from 'src/entities/memo.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository, InsertResult, UpdateResult, DeleteResult } from 'typeorm';
+import { MemoDTO } from './dto';
 // serviceでDBテーブルとのやりとりを行う
 @Injectable()
 export class MemoService {
@@ -13,21 +14,21 @@ export class MemoService {
   async list(): Promise<Memo[]> {
     return await this.memoRepository.find();
   }
-  async create(memo: {
-    title: string;
-    todo: string;
-    highPriority: boolean;
-  }): Promise<InsertResult> {
+  async create(memo: MemoDTO): Promise<InsertResult> {
     return await this.memoRepository.insert(memo);
   }
   async show(id: number): Promise<Memo | null> {
     return await this.memoRepository.findOne({ id: id });
   }
-  async update(id: number, memo): Promise<UpdateResult> {
+  async update(id: number, memo: MemoDTO): Promise<UpdateResult> {
     return await this.memoRepository.update(id, memo);
   }
 
   async delete(id: number): Promise<DeleteResult> {
+    const targetItem = await this.show(id);
+    if (!targetItem) {
+      return Promise.reject(new Error('memo not found.'));
+    }
     return await this.memoRepository.delete(id);
   }
 }
